+++ packages/eslint-plugin-jest/src/index.js
@@ -10,6 +10,7 @@
 import noDisabledTests from './rules/no_disabled_tests';
 import noFocusedTests from './rules/no_focused_tests';
 import noIdenticalTitle from './rules/no_identical_title';
+import noToBeFloat from './rules/no_to_be_float';
 import preferToHaveLength from './rules/prefer_to_have_length';
 import validExpect from './rules/valid_expect';
 
@@ -20,6 +21,7 @@ module.exports = {
         'jest/no-disabled-tests': 'warn',
         'jest/no-focused-tests': 'error',
         'jest/no-identical-title': 'error',
+        'jest/no-to-be-float': 'error',
         'jest/prefer-to-have-length': 'warn',
         'jest/valid-expect': 'error',
       },
@@ -52,6 +54,7 @@ module.exports = {
     'no-disabled-tests': noDisabledTests,
     'no-focused-tests': noFocusedTests,
     'no-identical-title': noIdenticalTitle,
+    'no-to-be-float': noToBeFloat,
     'prefer-to-have-length': preferToHaveLength,
     'valid-expect': validExpect,
   },